backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.29.7
cryptography==41.0.7
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
schedule==1.2.0
